requirements.txt
# Core dependencies
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
chromadb==1.0.15
sentence-transformers==5.0.0

# PDF processing
PyMuPDF==1.26.3
pdf2image==1.17.0
pytesseract==0.3.13
tabula-py==2.10.0

# Image processing
Pillow==10.1.0
opencv-python==4.12.0.88

# Data processing
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5

# ML/AI
torch==2.7.0
transformers==4.53.2
scikit-learn==1.5.1

# LLM and agents
openai==1.78.0

# Utilities
tqdm==4.66.5
python-dotenv==0.21.0
psutil==5.9.0  # Added for memory monitoring

# Storage and caching
diskcache==5.6.3
lmdb==1.4.1

# NLP
nltk==3.9.1

# Note: camelot-py is not compatible with Python 3.12
# You may need to use an alternative like pdfplumber for table extraction